requirements.txt
Django==1.11.6
django-crispy-forms==1.7.0
django-pagedown==0.1.3
djangorestframework==3.7.1
Markdown==2.6.9
mysqlclient==1.3.12
olefile==0.44
Pillow==4.3.0
pytz==2017.2
